test(mpv): add test suite for MPV command functionality
